+++ src/lib/sdkDappCore.ts
@@ -1,1 +1,0 @@
-export { stringIsInteger } from '@multiversx/sdk-dapp-core/out/utils/validation/stringIsInteger';