+++ src/pages/Faucet/components/FuacetContent/FaucetContent.tsx
@@ -1,6 +1,6 @@
-import { useRef, useState } from 'react';
+import { useEffect, useRef, useState } from 'react';
 import { Loader } from 'components';
-import { getEgldLabel } from 'lib';
+import { getEgldLabel, refreshAccount, useGetAccountInfo } from 'lib';
 import { DataTestIdsEnum } from 'localConstants';
 import {
   useGetFaucetSettingsQuery,
@@ -15,9 +15,15 @@ export const FaucetContent = () => {
   const [getFunds, { isSuccess }] = useRequestFundsMutation();
   const [fundsReceived, setFundsReceived] = useState(false);
   const [requestFailed, setRequestFailed] = useState('');
-
+  const { websocketEvent } = useGetAccountInfo();
   const { data: settings, error: settingsError } = useGetFaucetSettingsQuery();
   const egldLabel = getEgldLabel();
+
+  useEffect(() => {
+    if (isSuccess && fundsReceived) {
+      refreshAccount();
+    }
+  }, [isSuccess, fundsReceived, websocketEvent]);
 
   const handleRequestClick = async (captcha: string) => {
     const response = await getFunds(captcha);